Refactor to open separate connection to OTP redis
