+++ src/app.js
@@ -8,8 +8,8 @@ const port = process.env.TS_PORT || config.defaultPort;
 const app = express();
 const server = http.createServer(app);
 const podname = process.env.podname;
-// const redis = require("redis-promise");
-const redisOTP = require("./redis-otp");
+const redis = require("redis-promise");
+const redisOTP = require("./redis-otp/datastore");
 const gkeHostname = "ts-redis-master";
 const redisHost = process.env.NODE_ENV === "test" ? "127.0.0.1" : gkeHostname;
 const credentials = require("./credentials");
@@ -31,13 +31,13 @@ const start = ()=>{
 
     console.log(`server is listening on ${port}`);
 
-    //redis.initdb(null, redisHost);
-    redisOTP.init();
+    redis.initdb(null, redisHost);
+    redisOTP.initdb(null);
   })
 };
 
 const stop = ()=>{
-  redisOTP.close();
+  redis.close();
   server.close();
 };
